casext: refactor childDescriptors to MapDescriptors
